Refactor migrate typescript and handle imports
