+++ package.json
@@ -6,8 +6,8 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha",
-    "bot": "node src/bot.js",
-    "dev": "nodemon src/bot.js",
+    "bot": "npx tsx src/bot.ts",
+    "dev": "nodemon --exec npx tsx src/bot.ts",
     "summarizer": "node src/summarizer.js",
     "prettier": "npx prettier . --write"
   },
@@ -19,18 +19,18 @@
     "chai": "^5.1.1",
     "discord.js": "^14.14.1",
     "dotenv": "^16.4.5",
+    "eslint-import-resolver-typescript": "^3.6.1",
     "mocha": "^10.4.0",
-    "module-alias": "^2.2.3",
     "openai": "^4.31.0",
+    "path": "^0.12.7",
     "redis": "^4.6.13"
   },
   "devDependencies": {
     "nodemon": "^3.1.0",
-    "prettier": "3.2.5"
-  },
-  "_moduleAliases": {
-    "@lib": "src/lib",
-    "@config": "src/config.js",
-    "@prompts": "src/prompts.js"
+    "prettier": "3.2.5",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
+    "tsx": "^4.11.0",
+    "typescript": "^5.4.5"
   }
 }
