+++ src/commands/handleAutoPost.ts
