+++ src/commands/handleInfo.ts
