+++ src/commands/handleView.ts
