--- src/commands/index.js
@@ -1,19 +1,0 @@
-import commandBuilder from "./commandBuilder";
-import handleAutoPost from "./handleAutoPost";
-import handleCreate from "./handleCreate";
-import handleInfo from "./handleInfo";
-import handleModals from "./handleModals";
-import handleRespond from "./handleRespond";
-import handleRespondButton from "./handleRespondButton";
-import handleView from "./handleView";
-
-export {
-  commandBuilder,
-  handleAutoPost,
-  handleCreate,
-  handleInfo,
-  handleModals,
-  handleRespond,
-  handleRespondButton,
-  handleView,
-};
